Refactor and add tests
- Move code to find which user submitted a request to a method on ServiceXResource
- Split QueryTransformationRequest resource into two properly RESTful resources as follows
  - `GET /transformation` or `GET /transformation?submitted_by=<user_id>` to get a list of transformation requests
  - `GET /transformation/<request_id>` to get a single transformation request
- Add tests for both endpoints
