+++ servicex/resources/servicex_resource.py
@@ -26,17 +26,31 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 from datetime import datetime, timezone
+from typing import Optional
 
 from flask import current_app
+from flask_jwt_extended import get_jwt_identity
 from flask_restful import Resource
 
-from servicex.models import TransformationResult
+from servicex.models import TransformationResult, UserModel
 
 
 class ServiceXResource(Resource):
     @classmethod
     def _generate_advertised_endpoint(cls, endpoint):
         return "http://" + current_app.config['ADVERTISED_HOSTNAME'] + "/" + endpoint
+
+    @staticmethod
+    def get_requesting_user() -> Optional[UserModel]:
+        """
+        :return: User who submitted request for resource.
+        If auth is enabled, this cannot be None for JWT-protected resources
+        which are decorated with @auth_required or @admin_required.
+        """
+        user = None
+        if current_app.config.get('ENABLE_AUTH'):
+            user = UserModel.find_by_sub(get_jwt_identity())
+        return user
 
     @staticmethod
     def _generate_file_status_record(dataset_file, status):
