+++ tests/test_decorators.py
@@ -120,12 +120,13 @@ class TestDecorators(WebTestBase):
     def test_auth_decorator_integration_auth_disabled(self, mocker):
         fake_transform_id = 123
         data = {'id': fake_transform_id}
-        mock = mocker.patch('servicex.resources.query_transformation_request'
+        mock = mocker.patch('servicex.resources.transformation_request'
                             '.TransformRequest.return_request').return_value
         mock.to_json.return_value = data
         client = self._test_client(mocker)
         with client.application.app_context():
             response: Response = client.get(f'servicex/transformation/{fake_transform_id}')
+            print(response.data)
             assert response.status_code == 200
             assert response.json == data
 
@@ -157,7 +158,7 @@ class TestDecorators(WebTestBase):
         client = self._test_client(mocker, extra_config={'ENABLE_AUTH': True})
         fake_transform_id = 123
         data = {'id': fake_transform_id}
-        mock = mocker.patch('servicex.resources.query_transformation_request'
+        mock = mocker.patch('servicex.resources.transformation_request'
                             '.TransformRequest.return_request').return_value
         mock.submitted_by = user.id
         mock.to_json.return_value = data
